Added file to centralize task actions
