+++ packages/launchpad/src/index.ts
@@ -1,4 +1,5 @@
 export { default as Checklist } from './checklist';
 export { default as ChecklistItem } from './checklist-item';
 export { default as Launchpad } from './launchpad';
+export { setUpActions } from './setup-actions';
 export * from './types';
